chore: 폴더링 및 status, exception 설정
