+++ src/main/java/makar/dev/controller/UserController.java
@@ -1,0 +1,14 @@
+package makar.dev.controller;
+
+import lombok.RequiredArgsConstructor;
+import makar.dev.service.UserService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1")
+public class UserController {
+    private final UserService userService;
+
+}
