+++ src/main/java/makar/dev/converter/UserConverter.java
@@ -1,0 +1,4 @@
+package makar.dev.converter;
+
+public class UserConverter {
+}
