--- src/main/java/makar/dev/domain/Route/repository/RouteRepository.java
@@ -1,8 +1,0 @@
-package makar.dev.domain.Route.repository;
-
-
-import makar.dev.domain.Route.entity.Route;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RouteRepository extends JpaRepository<Route, Long> {
-}
