+++ src/main/java/makar/dev/domain/Station.java
@@ -1,0 +1,25 @@
+package makar.dev.domain;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Entity
+public class Station {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "station_id")
+    private Long stationId;
+
+    @Column(nullable = false)
+    private String stationName;
+
+    @ManyToOne
+    @JoinColumn(name="transfer_id", nullable = false)
+    private Transfer transfer;
+}
