--- src/main/java/makar/dev/domain/Station/repository/StationRepository.java
@@ -1,8 +1,0 @@
-package makar.dev.domain.Station.repository;
-
-
-import makar.dev.domain.Station.entity.Station;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-}
