--- src/main/java/makar/dev/domain/User/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package makar.dev.domain.User.repository;
-
-import makar.dev.domain.User.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
