+++ src/main/java/makar/dev/dto/request/UserRequest.java
@@ -1,0 +1,4 @@
+package makar.dev.dto.request;
+
+public class UserRequest {
+}
