+++ src/main/java/makar/dev/dto/response/RouteResponse.java
@@ -1,0 +1,4 @@
+package makar.dev.dto.response;
+
+public class RouteResponse {
+}
