+++ src/main/java/makar/dev/dto/response/StationResponse.java
@@ -1,0 +1,4 @@
+package makar.dev.dto.response;
+
+public class StationResponse {
+}
