+++ src/main/java/makar/dev/dto/response/UserResponse.java
@@ -1,0 +1,4 @@
+package makar.dev.dto.response;
+
+public class UserResponse {
+}
