+++ src/main/java/makar/dev/repository/RouteRepository.java
@@ -1,0 +1,8 @@
+package makar.dev.repository;
+
+
+import makar.dev.domain.Route;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RouteRepository extends JpaRepository<Route, Long> {
+}
