+++ src/main/java/makar/dev/repository/StationRepository.java
@@ -1,0 +1,8 @@
+package makar.dev.repository;
+
+
+import makar.dev.domain.Station;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StationRepository extends JpaRepository<Station, Long> {
+}
