+++ src/main/java/makar/dev/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package makar.dev.repository;
+
+import makar.dev.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
